+++ CHANGELOG.md
@@ -12,6 +12,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Breaking changes
 
+* In revsets, symbol expressions (such as change ID prefix) no longer resolve to
+  multiple revisions, and error out if resolved to more than one revisions. Use
+  `change_id(prefix)` or `bookmarks(exact:name)` to query divergent changes or
+  conflicted bookmarks.
+
 * `jj op abandon` now discards previous versions of a change (or predecessors)
   if they become unreachable from the operation history. The evolution history
   is truncated accordingly.